// src/generated/webhooks.ts
export type webhooks = Record<string, never>;
// ...
export interface operations {
    "webhooks.list": {
        parameters: {
            query?: {
                /** @description Filters the webhooks in the response by an `anchor_id` entity type. */
                anchor_type?: components["parameters"]["anchor_type"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that lists webhooks with webhook details. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["WebhookList"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "webhooks.post": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: {
            content: {
                "application/json": components["schemas"]["webhook"];
            };
        };
        responses: {
            /** @description A successful request returns the HTTP `201 Created` status code and a JSON response body with a [`webhook`](/docs/api/webhooks/v1/#definition-webhook) object that includes the webhook ID for later use. */
            201: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["webhook"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "webhooks.get": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description The ID of the webhook for which to list subscriptions. */
                webhook_id: components["parameters"]["webhook_id"];
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that shows webhook details. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["webhook"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "webhooks.delete": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description The ID of the webhook for which to list subscriptions. */
                webhook_id: components["parameters"]["webhook_id"];
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A successful request returns the HTTP `204 No Content` status code with no JSON response body. */
            204: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "webhooks.update": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description The ID of the webhook for which to list subscriptions. */
                webhook_id: components["parameters"]["webhook_id"];
            };
            cookie?: never;
        };
        requestBody?: {
            content: {
                "application/json": components["schemas"]["patch_request"];
            };
        };
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that shows webhook details. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["webhook"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "event-types.list": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description The ID of the webhook for which to list subscriptions. */
                webhook_id: components["parameters"]["webhook_id"];
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that lists event subscriptions for a webhook. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["EventTypeList"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
// ...
    "verify-webhook-signature.post": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: {
            content: {
                "application/json": components["schemas"]["verify_webhook_signature"];
            };
        };
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that shows the verification status. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["verify_webhook_signature_response"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "webhooks-event-types.list": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that lists available events to which any webhook can subscribe. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["EventTypeList"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
    "webhooks-events.list": {
        parameters: {
            query?: {
                /** @description The number of webhook event notifications to return in the response. */
                page_size?: components["parameters"]["page_size"];
                /** @description Filters the webhook event notifications in the response to those created on or after this date and time and on or before the `end_time` value. Both values are in [Internet date and time format](https://tools.ietf.org/html/rfc3339#section-5.6) format. Example: `start_time=2013-03-06T11:00:00Z`. */
                start_time?: components["parameters"]["start_time"];
                /** @description Filters the webhook event notifications in the response to those created on or after the `start_time` and on or before this date and time. Both values are in [Internet date and time format](https://tools.ietf.org/html/rfc3339#section-5.6) format. Example: `end_time=2013-03-06T11:00:00Z`. */
                end_time?: components["parameters"]["end_time"];
                /** @description Filters the response to a single transaction, by ID. */
                transaction_id?: components["parameters"]["transaction_id"];
                /** @description Filters the response to a single event. */
                event_type?: components["parameters"]["event_type"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A successful request returns the HTTP `200 OK` status code and a JSON response body that lists webhooks event notifications. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["EventList"];
                };
            };
            /** @description The error response. */
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["error-2"];
                };
            };
        };
    };
// ...
}